Moved navbar component to navbar directory
