+++ frontend/src/components/navbar/Navbar.scss
@@ -1,4 +1,4 @@
-@use '../styles/variables.scss' as *;
+@use '../../styles/variables.scss' as *;
 
 $desktop-font-size: 1.4rem;
 $desktop-font-size2: 1rem;
