+++ frontend/src/pages/Home/Home.jsx
@@ -1,4 +1,4 @@
-import Navbar from '../../components/Navbar.jsx';
+import Navbar from '../../components/navbar/Navbar.jsx';
 
 export default function Home() {
   return (
